+++ gitlint/options.py
@@ -4,6 +4,18 @@ import os
 from gitlint.utils import ustr, sstr
 
 
+def allow_none(func):
+    """ Decorator that sets option value to None if the passed value is None, otherwise calls the regular set method """
+
+    def wrapped(obj, value):
+        if value is None:
+            obj.value = None
+        else:
+            func(obj, value)
+
+    return wrapped
+
+
 class RuleOptionError(Exception):
     pass
 
@@ -43,6 +55,7 @@ class RuleOption(object):
 
 
 class StrOption(RuleOption):
+    @allow_none
     def set(self, value):
         self.value = ustr(value)
 
@@ -59,6 +72,7 @@ class IntOption(RuleOption):
             error_msg = u"Option '{0}' must be a positive integer (current value: '{1}')".format(self.name, value)
         raise RuleOptionError(error_msg)
 
+    @allow_none
     def set(self, value):
         try:
             self.value = int(value)
@@ -70,6 +84,8 @@ class IntOption(RuleOption):
 
 
 class BoolOption(RuleOption):
+
+    # explicit choice to not annotate with @allow_none: Booleans must be False or True, they cannot be unset.
     def set(self, value):
         value = ustr(value).strip().lower()
         if value not in ['true', 'false']:
@@ -81,6 +97,7 @@ class ListOption(RuleOption):
     """ Option that is either a given list or a comma-separated string that can be splitted into a list when being set.
     """
 
+    @allow_none
     def set(self, value):
         if isinstance(value, list):
             the_list = value
@@ -97,6 +114,7 @@ class PathOption(RuleOption):
         self.type = type
         super(PathOption, self).__init__(name, value, description)
 
+    @allow_none
     def set(self, value):
         value = ustr(value)